requirements.txt
# =========================
# LangChain Core
# =========================
langchain==0.3.0
langchain-community==0.3.0
langchain-experimental==0.3.0


# LLM Providers
langchain-openai==0.2.0
langchain-ollama==0.2.0
langchain-cohere==0.3.0
langchain-groq==0.2.0

# =========================
# Vector DB
# =========================
langchain-qdrant==0.1.4
qdrant-client==1.11.0
chromadb==0.5.15

# =========================
# Retrieval & Evaluation
# =========================
rank-bm25==0.2.2
ragas==0.2.0   # ✅ FIXED VERSION

# =========================
# Document Processing
# =========================
pypdf==4.3.1
unstructured==0.15.13
python-docx==1.1.0

# =========================
# API / UI
# =========================
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0

# =========================
# Core Utilities
# =========================

beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.2

# =========================
# LLM SDKs
# =========================
requests==2.31.0
urllib3==1.26.18
charset_normalizer==2.0.12

cohere==5.11.0

# =========================
# Supporting libs
# =========================
tiktoken==0.7.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
httpx==0.27.2

# =========================
# Dev
# =========================
ipykernel==6.29.5
jupyter==1.0.0